refactoring to avoid circular import
